+++ src/Ouzo/Utilities/Joiner.php
@@ -6,6 +6,7 @@ class Joiner
     private $_separator;
     private $_skipNulls;
     private $_function;
+    private $_valuesFunction;
 
     public function __construct($separator)
     {
@@ -24,10 +25,14 @@ class Joiner
     public function join(array $array)
     {
         $function = $this->_function;
+        $valuesFunction = $this->_valuesFunction;
         $result = '';
         foreach ($array as $key => $value) {
             if (!$this->_skipNulls || ($this->_skipNulls && $value)) {
-                $result .= ($function ? $function($key, $value) : $value) . $this->_separator;
+                $result .= (
+                    $function ? $function($key, $value) :
+                        ($valuesFunction ? $valuesFunction($value) : $value)
+                    ) . $this->_separator;
             }
         }
         return rtrim($result, $this->_separator);
@@ -45,6 +50,12 @@ class Joiner
     public function map($function)
     {
         $this->_function = $function;
+        return $this;
+    }
+
+    public function mapValues($function)
+    {
+        $this->_valuesFunction = $function;
         return $this;
     }
 }